--- cpu/cortex-m0_common/Makefile
@@ -1,4 +1,0 @@
-# define the module that is build
-MODULE = cortex-m0_common
-
-include $(RIOTBASE)/Makefile.base
