--- cpu/cortex-m3_common/Makefile.include
@@ -1,3 +1,0 @@
-
-# include module specific includes
-export INCLUDES += -I$(RIOTCPU)/cortex-m3_common/include
