--- cpu/cortex-m4_common/Makefile
@@ -1,4 +1,0 @@
-# define the module that is build
-MODULE = cortex-m4_common
-
-include $(RIOTBASE)/Makefile.base
