+++ cpu/cortexm_common/atomic_arch.c
@@ -1,13 +1,13 @@
 /*
- * Copyright (C) 2014 Freie Universität Berlin
+ * Copyright (C) 2014-2015 Freie Universität Berlin
  *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
+ * This file is subject to the terms and conditions of the GNU Lesser General
+ * Public License v2.1. See the file LICENSE in the top level directory for more
+ * details.
  */
 
 /**
- * @ingroup     cpu_cortexm3_common
+ * @ingroup     cpu_cortexm_common
  * @{
  *
  * @file
@@ -25,6 +25,7 @@
 #include "irq.h"
 #include "cpu.h"
 
+#if ARCH_HAS_ATOMIC_COMPARE_AND_SWAP
 int atomic_cas(atomic_int_t *var, int old, int now)
 {
     int tmp;
@@ -44,3 +45,4 @@ int atomic_cas(atomic_int_t *var, int old, int now)
 
     return (status == 0);
 }
+#endif
