+++ cpu/cortexm_common/include/core_cmFunc.h
