+++ cpu/cortexm_common/include/core_cmSimd.h
