+++ cpu/cortexm_common/include/cpu.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014 Freie Universität Berlin
+ * Copyright (C) 2014-2015 Freie Universität Berlin
  *
  * This file is subject to the terms and conditions of the GNU Lesser
  * General Public License v2.1. See the file LICENSE in the top level
@@ -7,16 +7,18 @@
  */
 
 /**
- * @defgroup    cpu_cortexm3_common ARM Cortex-M3 common
+ * @defgroup    cpu_cortexm_common ARM Cortex-M common
  * @ingroup     cpu
- * @brief       Common implementations and headers for Cortex-M3 family based micro-controllers
+ * @brief       Common implementations and headers for Cortex-M family based
+ *              micro-controllers
  * @{
  *
  * @file
- * @brief       Basic definitions for the Cortex-M3 common module
+ * @brief       Basic definitions for the Cortex-M common module
  *
- * When ever you want to do something hardware related, that is accessing MCUs registers directly,
- * just include this file. It will then make sure that the MCU specific headers are included.
+ * When ever you want to do something hardware related, that is accessing MCUs
+ * registers, just include this file. It will then make sure that the MCU
+ * specific headers are included.
  *
  * @author      Stefan Pfeiffer <stefan.pfeiffer@fu-berlin.de>
  * @author      Hauke Petersen <hauke.petersen@fu-berlin.de>
@@ -26,15 +28,10 @@
 #ifndef CORTEXM_COMMON_H_
 #define CORTEXM_COMMON_H_
 
-/**
- * @brief Cortex-M3 has architecture specific atomic operations in atomic_arch.c.
- */
-#define ARCH_HAS_ATOMIC_COMPARE_AND_SWAP 1
 
 #include "cpu_conf.h"
 
 /**
- * For downwards compatibility with old RIOT code.
  * TODO: remove once core was adjusted
  */
 #include "irq.h"
@@ -52,7 +49,16 @@ extern "C" {
 /** @} */
 
 /**
- * @brief Initialization of the CPU
+ * @brief   Some members of the Cortex-M family have architecture specific atomic
+ *          operations in atomic_arch.c
+ */
+#if defined(CPU_ARCH_CORTEX_M3) || defined(CPU_ARCH_CORTEX_M4) || \
+    defined(CPU_ARCH_CORTEX_M4F)
+#define ARCH_HAS_ATOMIC_COMPARE_AND_SWAP 1
+#endif
+
+/**
+ * @brief   Initialization of the CPU
  */
 void cpu_init(void);
